Restructure cDAC data generation.
