+++ src/coreclr/debug/datadescriptor-shared/contract-descriptor.c.in
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-#include <stdint.h>
 #include "contractconfiguration.h"
 
 #ifdef _MSC_VER
@@ -9,17 +8,6 @@
 #else
 #define DLLEXPORT __attribute__((visibility("default")))
 #endif
-
-struct ContractDescriptor
-{
-    uint64_t magic;
-    uint32_t flags;
-    const uint32_t descriptor_size;
-    const char *descriptor;
-    const uint32_t pointer_data_count;
-    uint32_t pad0;
-    const void** pointer_data;
-};
 
 // POINTER_DATA_NAME and CONTRACT_NAME are macros provided by
 // contractconfiguration.h which is configured by CMake
