+++ src/coreclr/debug/datadescriptor-shared/contractconfiguration.h.in
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <contract-descriptor.h>
+
 #define POINTER_DATA_NAME @POINTER_DATA_NAME@
 #define CONTRACT_NAME @CONTRACT_NAME@
