+++ src/coreclr/debug/datadescriptor-shared/contractdescriptorstub.c
@@ -10,17 +10,6 @@
 #define DLLEXPORT __attribute__((visibility("default")))
 #endif
 
-struct ContractDescriptor
-{
-    uint64_t magic;
-    uint32_t flags;
-    const uint32_t descriptor_size;
-    const char *descriptor;
-    const uint32_t pointer_data_count;
-    uint32_t pad0;
-    const void** pointer_data;
-};
-
 // POINTER_DATA_NAME and CONTRACT_NAME are macros provided by
 // contractconfiguration.h which is configured by CMake
 extern const void* POINTER_DATA_NAME[];
