+++ src/coreclr/interpreter/CMakeLists.txt
@@ -16,11 +16,11 @@ set(INTERPRETER_LINK_LIBRARIES
 )
 
 if(CLR_CMAKE_HOST_WIN32)
-    list(APPEND INTERPRETER_LINK_LIBRARIES
-        ${STATIC_MT_CRT_LIB}
-        ${STATIC_MT_VCRT_LIB}
-    )
-endif(CLR_CMAKE_HOST_WIN32)
+  list(APPEND INTERPRETER_LINK_LIBRARIES
+    ${STATIC_MT_CRT_LIB}
+    ${STATIC_MT_VCRT_LIB}
+  )
+endif()
 
 if(CLR_CMAKE_HOST_WIN32)
     set(CLRINTERPRETER_EXPORTS ${CMAKE_CURRENT_LIST_DIR}/clrinterpreter.exports)
